missing file

+++ gridftp/client/source/globus_ftp_client_feature.c
@@ -372,7 +372,7 @@ globus_ftp_client_is_feature_supported(
         error = GLOBUS_I_FTP_CLIENT_ERROR_NULL_PARAMETER("u_features");
         goto error_param;
     }
-    if(feature < 0 || feature >= GLOBUS_FTP_CLIENT_FEATURE_MAX)
+    if(((int) feature < 0) || feature >= GLOBUS_FTP_CLIENT_FEATURE_MAX)
     {
         error = GLOBUS_I_FTP_CLIENT_ERROR_INVALID_PARAMETER("feature");
         goto error_param;
